refactor(models): lift shared SpatialModel/SpatialPanelModel API into _SpatialModelBase
Phase 4a of the panel/cross-section unification: extract methods that
were byte-identical (or only differed in docstring wording) between
SpatialModel and SpatialPanelModel into a new abstract parent class
_SpatialModelBase in bayespecon/models/_common.py.

Lifted: fit, summary, inference_data, pymc_model, fitted_values,
residuals, spatial_effects, _nonintercept_indices,
_nonintercept_feature_names, _spatial_lag_column_indices, _add_nu_prior,
_beta_names, _model_coords, _rename_summary_index,
_attach_jacobian_corrected_log_likelihood, _lazy_lm_test,
_run_lm_diagnostics, _require_fit, _posterior_mean,
_spatial_diagnostics_tests, and the abstract hooks _build_pymc_model,
_compute_spatial_effects, _compute_spatial_effects_posterior,
_fitted_mean_from_posterior.

Kept divergent surface (different prefixes, different W parsing) on the
two subclasses for now — to be unified in Phase 4b.

Net: -476 LOC. 957/957 tests pass.
