+++ bayespecon/models/panel_base.py
@@ -3,7 +3,6 @@
 from __future__ import annotations
 
 import warnings
-from abc import ABC, abstractmethod
 from typing import Any, Optional, Union
 
 import arviz as az
@@ -21,7 +20,7 @@ from ..logdet import (
     make_logdet_numpy_fn,
     make_logdet_numpy_vec_fn,
 )
-from ._sampler import prepare_compile_kwargs, prepare_idata_kwargs
+from ._common import _SpatialModelBase
 from .base import _is_row_standardized_csr
 
 
@@ -210,7 +209,7 @@ def _parse_panel_W(
     return W_csr, row_std
 
 
-class SpatialPanelModel(ABC):
+class SpatialPanelModel(_SpatialModelBase):
     """Base class for static spatial panel models with FE transforms.
 
     Holds the within-transformation, panel-aware sorting, and weights
@@ -700,380 +699,6 @@ class SpatialPanelModel(ABC):
 
         return _chunked_eig_means(rho_draws, eigs, weights=eigs * V_col_sums * c)
 
-    @property
-    def _nonintercept_indices(self) -> list[int]:
-        """Return indices of non-constant (non-intercept) columns in X.
-
-        This is used to exclude the intercept from impact measures, since
-        the intercept has no meaningful spatial effect interpretation.
-
-        Returns
-        -------
-        list[int]
-            Column indices of X that are not constant/intercept columns.
-        """
-        indices: list[int] = []
-        for j, name in enumerate(self._feature_names):
-            column = self._X[:, j]
-            is_named_intercept = name.lower() == "intercept"
-            is_constant = np.allclose(column, column[0])
-            if not (is_named_intercept or is_constant):
-                indices.append(j)
-        return indices
-
-    @property
-    def _nonintercept_feature_names(self) -> list[str]:
-        """Return feature names for non-intercept columns.
-
-        Returns
-        -------
-        list[str]
-            Feature names excluding intercept/constant columns.
-        """
-        return [self._feature_names[i] for i in self._nonintercept_indices]
-
-    @staticmethod
-    def _spatial_lag_column_indices(
-        X: np.ndarray, feature_names: list[str]
-    ) -> list[int]:
-        """Return indices of regressors that should receive spatial lags.
-
-        Constant columns are treated as intercept-like and excluded, which
-        avoids adding redundant ``W * intercept`` terms to SLX/Durbin models.
-
-        Parameters
-        ----------
-        X : np.ndarray
-            Raw panel design matrix before FE transformation.
-        feature_names : list[str]
-            Column labels aligned with ``X``.
-
-        Returns
-        -------
-        list[int]
-            Column indices eligible for spatial lags.
-        """
-        indices: list[int] = []
-        for j, name in enumerate(feature_names):
-            column = X[:, j]
-            is_named_intercept = name.lower() == "intercept"
-            is_constant = np.allclose(column, column[0])
-            if not (is_named_intercept or is_constant):
-                indices.append(j)
-        return indices
-
-    def _add_nu_prior(self, model: pm.Model) -> pm.Model:
-        """Add the degrees-of-freedom prior for robust (Student-t) models.
-
-        Called inside ``_build_pymc_model`` when ``self.robust`` is True.
-        Uses an :math:`\\mathrm{Exp}(\\lambda_\\nu)` prior on ``nu`` with rate ``nu_lam`` (default
-        1/30, giving mean ≈ 30, favouring near-Normal tails). A lower
-        bound of 2 is enforced so that the variance exists.
-
-        Parameters
-        ----------
-        model : pymc.Model
-            The model context in which to add the ``nu`` prior.
-
-        Returns
-        -------
-        pymc.Model
-            The same model context (``nu`` is added as a side effect).
-        """
-        nu_lam = self.priors.get("nu_lam", 1.0 / 30.0)
-        pm.Truncated("nu", pm.Exponential.dist(lam=nu_lam), lower=2.0)
-        return model
-
-    @abstractmethod
-    def _build_pymc_model(self) -> pm.Model:
-        """Construct and return a pm.Model."""
-
-    @abstractmethod
-    def _compute_spatial_effects(self) -> dict[str, np.ndarray]:
-        """Compute direct/indirect/total effects at posterior mean."""
-
-    @abstractmethod
-    def _compute_spatial_effects_posterior(
-        self,
-    ) -> tuple[np.ndarray, np.ndarray, np.ndarray]:
-        """Compute direct, indirect, and total effects for each posterior draw.
-
-        Returns
-        -------
-        tuple of np.ndarray
-            ``(direct_samples, indirect_samples, total_samples)`` where each
-            array has shape ``(G, k)`` or ``(G, k_wx)``.
-        """
-
-    @abstractmethod
-    def _fitted_mean_from_posterior(self) -> np.ndarray:
-        """Posterior-mean fitted values on transformed scale."""
-
-    def _beta_names(self) -> list[str]:
-        """Return coefficient labels used for posterior summaries.
-
-        Returns
-        -------
-        list[str]
-            Coefficient labels aligned with the ``beta`` parameter.
-        """
-        return list(self._feature_names)
-
-    def _model_coords(self) -> dict[str, list[str]]:
-        """Return PyMC coordinate labels for named dimensions.
-
-        Returns
-        -------
-        dict[str, list[str]]
-            Coordinates passed to :class:`pymc.Model`.
-        """
-        return {"coefficient": self._beta_names()}
-
-    @staticmethod
-    def _rename_summary_index(summary_df: pd.DataFrame) -> pd.DataFrame:
-        """Strip the ``beta[...]`` wrapper from coefficient row labels.
-
-        Parameters
-        ----------
-        summary_df : pandas.DataFrame
-            ArviZ summary output.
-
-        Returns
-        -------
-        pandas.DataFrame
-            Summary with human-readable coefficient row labels.
-        """
-        renamed = []
-        for label in summary_df.index.astype(str):
-            if label.startswith("beta[") and label.endswith("]"):
-                renamed.append(label[5:-1])
-            else:
-                renamed.append(label)
-        out = summary_df.copy()
-        out.index = renamed
-        return out
-
-    def fit(
-        self,
-        draws: int = 2000,
-        tune: int = 1000,
-        chains: int = 4,
-        target_accept: float = 0.9,
-        random_seed: Optional[int] = None,
-        **sample_kwargs,
-    ) -> az.InferenceData:
-        """Sample the posterior for the panel model.
-
-        Parameters
-        ----------
-        draws : int, default=2000
-            Number of post-tuning draws per chain.
-        tune : int, default=1000
-            Number of tuning draws per chain.
-        chains : int, default=4
-            Number of chains.
-        target_accept : float, default=0.9
-            NUTS target acceptance probability.
-        random_seed : int, optional
-            Random seed used by PyMC.
-        **sample_kwargs
-            Extra keyword arguments forwarded to :func:`pymc.sample`.  Pass
-            ``nuts_sampler="blackjax"`` (or ``"numpyro"``, ``"nutpie"``) to
-            select an alternative NUTS backend; defaults to PyMC's built-in
-            sampler.
-
-        Returns
-        -------
-        arviz.InferenceData
-            Posterior samples and diagnostics.
-        """
-        nuts_sampler = sample_kwargs.pop("nuts_sampler", "pymc")
-        try:
-            model = self._build_pymc_model(nuts_sampler=nuts_sampler)
-        except TypeError:
-            # Subclasses that don't accept ``nuts_sampler`` build the same
-            # model on every backend.
-            model = self._build_pymc_model()
-        self._pymc_model = model
-        if "idata_kwargs" in sample_kwargs:
-            sample_kwargs["idata_kwargs"] = prepare_idata_kwargs(
-                sample_kwargs["idata_kwargs"], model, nuts_sampler
-            )
-        sample_kwargs = prepare_compile_kwargs(sample_kwargs, nuts_sampler)
-        with model:
-            self._idata = pm.sample(
-                draws=draws,
-                tune=tune,
-                chains=chains,
-                target_accept=target_accept,
-                random_seed=random_seed,
-                nuts_sampler=nuts_sampler,
-                **sample_kwargs,
-            )
-        return self._idata
-
-    def _attach_jacobian_corrected_log_likelihood(
-        self,
-        idata: az.InferenceData,
-        spatial_param: str,
-        T: int = 1,
-    ) -> None:
-        """Add Jacobian correction to the auto-captured log-likelihood group.
-
-        For models that use ``pm.Normal("obs", observed=y)`` plus
-        ``pm.Potential("jacobian", logdet_fn(rho))``, PyMC auto-captures
-        the Gaussian part in the ``log_likelihood`` group but the Jacobian
-        term is absent.  This method adds the per-observation Jacobian
-        contribution ``log|I - ρW| * T / n`` to each pointwise LL value.
-
-        Notes
-        -----
-        The full panel log-determinant of the spatial filter is
-        :math:`T \\log|I_N - \\rho W|` because the stacked
-        :math:`(N T) \\times (N T)` filter is
-        :math:`I_T \\otimes (I_N - \\rho W)`, whose determinant is
-        :math:`|I_N - \\rho W|^T` by the Kronecker product rule.  Dividing
-        by :math:`n = N T` distributes that scalar Jacobian uniformly
-        over the per-observation pointwise log-likelihood entries that
-        ArviZ expects, so quantities like LOO and WAIC are computed on a
-        per-observation log density.  For dynamic panels the time
-        dimension is ``T - 1`` (one period is consumed by the lag), so
-        callers pass ``T = T - 1`` here.
-
-        Parameters
-        ----------
-        idata : arviz.InferenceData
-            InferenceData with an existing ``log_likelihood`` group.
-        spatial_param : str
-            Name of the spatial autoregressive parameter (``"rho"`` or
-            ``"lam"``) in the posterior.
-        T : int, default 1
-            Panel time-period multiplier for the Jacobian.
-        """
-        import xarray as xr
-
-        if "log_likelihood" not in idata.groups():
-            return
-
-        n = self._y.shape[0]
-        param_draws = idata.posterior[spatial_param].values.reshape(-1)  # (n_draws,)
-
-        # Jacobian: log|I - param*W| * T (pure numpy, respects logdet_method)
-        jacobian = self._logdet_numpy_vec_fn(param_draws) * T  # (n_draws,)
-        ll_jac = jacobian[:, None] / n  # (n_draws, 1)
-
-        # Add Jacobian to each variable in the log_likelihood group
-        n_chains = idata.posterior.sizes["chain"]
-        n_draws_per_chain = idata.posterior.sizes["draw"]
-        ll_jac_3d = ll_jac.reshape(n_chains, n_draws_per_chain, 1)  # broadcast over obs
-
-        new_vars = {}
-        for var_name in list(idata.log_likelihood.data_vars):
-            da = idata.log_likelihood[var_name]
-            # Use numpy addition + broadcast to avoid xarray alignment issues
-            # when the observation dimension name differs (e.g., "obs_dim_0" vs "obs_dim")
-            new_vals = da.values + ll_jac_3d
-            new_vars[var_name] = xr.DataArray(
-                new_vals,
-                dims=da.dims,
-                coords={k: v for k, v in da.coords.items() if k != da.dims[-1]},
-            )
-
-        idata["log_likelihood"] = xr.Dataset(new_vars)
-
-    @property
-    def inference_data(self) -> Optional[az.InferenceData]:
-        """Return the ArviZ InferenceData from the most recent fit.
-
-        Returns
-        -------
-        arviz.InferenceData or None
-            The inference data object, or ``None`` if the model has not
-            been fit yet.
-        """
-        return self._idata
-
-    @property
-    def pymc_model(self) -> Optional[pm.Model]:
-        """Return the PyMC model object built for the most recent fit.
-
-        Returns
-        -------
-        pymc.Model or None
-            The model object used by :meth:`fit`, or ``None`` if the instance
-            has not been fit yet.
-        """
-        return self._pymc_model
-
-    def _require_fit(self):
-        if self._idata is None:
-            raise RuntimeError("Model has not been fit yet. Call .fit() first.")
-
-    def _posterior_mean(self, var: str) -> np.ndarray:
-        return self._idata.posterior[var].mean(("chain", "draw")).to_numpy()
-
-    def summary(self, var_names: Optional[list] = None, **kwargs) -> pd.DataFrame:
-        """Return posterior summary table.
-
-        Parameters
-        ----------
-        var_names : list, optional
-            Variable names to include.
-        **kwargs
-            Additional arguments passed to :func:`arviz.summary`.
-
-        Returns
-        -------
-        pandas.DataFrame
-            Posterior summary table.
-        """
-        self._require_fit()
-        summary_df = az.summary(self._idata, var_names=var_names, **kwargs)
-        return self._rename_summary_index(summary_df)
-
-    def fitted_values(self) -> np.ndarray:
-        """Return fitted values at posterior mean parameters.
-
-        Returns
-        -------
-        np.ndarray
-            Fitted values on transformed panel scale.
-        """
-        self._require_fit()
-        return self._fitted_mean_from_posterior()
-
-    def residuals(self) -> np.ndarray:
-        """Return transformed residuals ``y - fitted``.
-
-        Returns
-        -------
-        np.ndarray
-            Residual vector on transformed panel scale.
-        """
-        self._require_fit()
-        return self._y - self.fitted_values()
-
-    # ------------------------------------------------------------------
-    # Class-level registry of applicable Bayesian LM specification tests.
-    # ------------------------------------------------------------------
-
-    @staticmethod
-    def _lazy_lm_test(module: str, name: str):
-        """Return a callable that lazily imports ``name`` from ``module``.
-
-        Used in ``_spatial_diagnostics_tests`` registries to avoid
-        circular imports at module-load time.
-        """
-        import importlib
-
-        def _fn(model):
-            mod = importlib.import_module(module)
-            return getattr(mod, name)(model)
-
-        return _fn
-
-    _spatial_diagnostics_tests: list[tuple] = []
-
     def spatial_diagnostics(self) -> pd.DataFrame:
         """Run Bayesian LM specification tests and return a summary table.
 
@@ -1217,80 +842,6 @@ class SpatialPanelModel(ABC):
             fmt=format,
             title=f"{model_type} decision tree (alpha={alpha})",
         )
-
-    def spatial_effects(
-        self, return_posterior_samples: bool = False
-    ) -> "pd.DataFrame | tuple[pd.DataFrame, dict[str, np.ndarray]]":
-        """Compute Bayesian inference for direct, indirect, and total impacts.
-
-        Computes impact measures for each posterior draw, then summarises
-        the posterior distribution with means, 95% credible intervals, and
-        Bayesian p-values.
-
-        Parameters
-        ----------
-        return_posterior_samples : bool, optional
-            If ``True``, return a ``(DataFrame, dict)`` tuple where the
-            dict contains the full posterior draws under keys
-            ``"direct"``, ``"indirect"``, and ``"total"``.  Default
-            ``False``.
-
-        Returns
-        -------
-        pd.DataFrame or tuple of (pd.DataFrame, dict)
-            If *return_posterior_samples* is ``False`` (default), returns
-            a DataFrame indexed by feature names with columns for posterior
-            means, credible-interval bounds, and Bayesian p-values.
-
-            If *return_posterior_samples* is ``True``, returns
-            ``(DataFrame, dict)`` where the dict has keys
-            ``"direct"``, ``"indirect"``, ``"total"``, each mapping
-            to a ``(G, k)`` array of posterior draws.
-        """
-        from ..diagnostics.spatial_effects import _build_effects_dataframe
-
-        self._require_fit()
-        direct_samples, indirect_samples, total_samples = (
-            self._compute_spatial_effects_posterior()
-        )
-
-        # Determine feature names based on the shape of the posterior samples.
-        # Models with WX terms (SDM, SLX, SDEM) report effects only for
-        # lagged covariates (k_wx columns), while models without WX terms
-        # (SAR, SEM) report effects for non-intercept covariates.
-        k_effects = direct_samples.shape[1]
-        if (
-            hasattr(self, "_wx_feature_names")
-            and len(self._wx_feature_names) == k_effects
-        ):
-            feature_names = list(self._wx_feature_names)
-        elif (
-            hasattr(self, "_nonintercept_feature_names")
-            and len(self._nonintercept_feature_names) == k_effects
-        ):
-            feature_names = list(self._nonintercept_feature_names)
-        else:
-            feature_names = list(self._feature_names[:k_effects])
-
-        # Determine model type label
-        model_type = self.__class__.__name__
-
-        df = _build_effects_dataframe(
-            direct_samples=direct_samples,
-            indirect_samples=indirect_samples,
-            total_samples=total_samples,
-            feature_names=feature_names,
-            model_type=model_type,
-        )
-
-        if return_posterior_samples:
-            posterior_samples = {
-                "direct": direct_samples,
-                "indirect": indirect_samples,
-                "total": total_samples,
-            }
-            return df, posterior_samples
-        return df
 
     def __repr__(self) -> str:
         n, k = self._X.shape
